+++ test/e2e/deploy_native_test.go
@@ -26,17 +26,8 @@ func TestDeploymentOfNativeContract(t *testing.T) {
 
 	counterStart := uint64(100 * rand.Intn(1000))
 
-	// transaction to deploy the contract
-	deploy := builders.Transaction().
-		WithMethod("_Deployments", "deployService").
-		WithArgs(
-			fmt.Sprintf("CounterFrom%d", counterStart),
-			uint32(protocol.PROCESSOR_TYPE_NATIVE),
-			[]byte(contracts.NativeSourceCodeForCounter(counterStart)),
-		).Builder()
-
 	printTestTime(t, "send deploy - start", &lt)
-	response, err := h.sendTransaction(deploy)
+	response, err := h.deployNativeContract(fmt.Sprintf("CounterFrom%d", counterStart), []byte(contracts.NativeSourceCodeForCounter(counterStart)))
 	printTestTime(t, "send deploy - end", &lt)
 
 	require.NoError(t, err, "deploy transaction should not return error")
